fix(deprecation): reflect staged deprecation in workspace UI version-history
The bit log staged-deprecation overlay lived only in ComponentLogMain.getLogs (CLI). The UI fetches logs via component.getLogs -> Workspace.getLogs, which bypassed it, so staged range-deprecation wasn't shown until tag/snap. Move the overlay into the shared Workspace.getLogs and route ComponentLogMain through it.
